# src/main/java/io/github/yuokada/npb/PlayerResource.java
package io.github.yuokada.npb;

import io.github.yuokada.npb.model.ErrorMessage;
import jakarta.enterprise.context.ApplicationScoped;
import jakarta.inject.Inject;
import jakarta.transaction.Transactional;
import jakarta.ws.rs.BadRequestException;
import jakarta.ws.rs.Consumes;
import jakarta.ws.rs.DELETE;
import jakarta.ws.rs.GET;
import jakarta.ws.rs.NotFoundException;
import jakarta.ws.rs.POST;
import jakarta.ws.rs.PUT;
import jakarta.ws.rs.Path;
import jakarta.ws.rs.PathParam;
import jakarta.ws.rs.Produces;
import jakarta.ws.rs.QueryParam;
import jakarta.ws.rs.core.MediaType;
import jakarta.ws.rs.core.Response;
import jakarta.ws.rs.core.Response.Status;
import java.time.LocalDateTime;
import java.time.format.DateTimeParseException;
import java.util.List;
import java.util.Set;
import org.eclipse.microprofile.openapi.annotations.media.Content;
import org.eclipse.microprofile.openapi.annotations.media.Schema;
import org.eclipse.microprofile.openapi.annotations.responses.APIResponse;
import org.eclipse.microprofile.openapi.annotations.responses.APIResponseSchema;
import org.eclipse.microprofile.openapi.annotations.responses.APIResponses;

@Path("/v1/player")
@Consumes(MediaType.APPLICATION_JSON)
@Produces(MediaType.APPLICATION_JSON)
@ApplicationScoped
public class PlayerResource {

    private static final Set<String> SUPPORTED_POSITIONS = Set.of("P", "C", "IF", "OF");
    private static final int PLAYER_NAME_MAX_LENGTH = 64;
    private static final int MIN_UNIFORM_NUMBER = 0;
    private static final int MAX_UNIFORM_NUMBER = 999;
    private static final int DEFAULT_TRANSFER_HISTORY_LIMIT = 100;
    private static final int MAX_TRANSFER_HISTORY_LIMIT = 500;

    @Inject
    PlayerRepository repository;

    @Inject
    TeamRepository teamRepository;

    @Inject
    PlayerTransferHistoryRepository transferHistoryRepository;

    @GET
    @Path("/{id}")
    @APIResponseSchema(value = Player.class, responseCode = "200")
    public Response getById(@PathParam("id") Integer playerId) {
        Player player = repository.findByIdOptional(playerId).orElseThrow(NotFoundException::new);
        return Response.ok(player).build();
    }

    @GET
    @Path("/{id}/transfers")
    @APIResponseSchema(value = PlayerTransferHistoryResponse[].class, responseCode = "200")
    public Response getTransfers(
        @PathParam("id") Integer playerId,
        @QueryParam("from") String from,
        @QueryParam("to") String to,
        @QueryParam("limit") Integer limit
    ) {
        repository.findByIdOptional(playerId).orElseThrow(NotFoundException::new);
        LocalDateTime fromDateTime = parseDateTimeOrNull(from, "from");
        LocalDateTime toDateTime = parseDateTimeOrNull(to, "to");
        int validatedLimit = validateTransferHistoryLimit(limit);

        if (fromDateTime != null && toDateTime != null && fromDateTime.isAfter(toDateTime)) {
            throw new BadRequestException("from must be less than or equal to to.");
        }

        List<PlayerTransferHistoryResponse> responses = transferHistoryRepository
            .search(playerId, fromDateTime, toDateTime, validatedLimit)
            .stream()
            .map(history -> new PlayerTransferHistoryResponse(
                history.id,
                history.player.id,
                history.fromTeam == null ? null : history.fromTeam.id,
                history.toTeam.id,
                history.transferredAt,
                history.createdAt
            ))
            .toList();
        return Response.ok(responses).build();
    }

    @POST
    @Transactional
    @Path("/{id}/transfers")
    @APIResponses({
        @APIResponse(responseCode = "200"),
        @APIResponse(responseCode = "400"),
        @APIResponse(responseCode = "404"),
        @APIResponse(responseCode = "409", content = @Content(schema = @Schema(implementation = ErrorMessage.class))),
        @APIResponse(responseCode = "500")
    })
    public Response transfer(@PathParam("id") Integer playerId, PlayerTransferRequest request) {
        if (request == null || request.toTeamId() == null) {
            throw new BadRequestException("toTeamId is required.");
        }

        Player player = repository.findByIdOptional(playerId).orElseThrow(NotFoundException::new);
        Team fromTeam = player.team;
        Team toTeam = teamRepository.findByIdOptional(request.toTeamId()).orElseThrow(NotFoundException::new);
        if (fromTeam.id.equals(toTeam.id)) {
            throw new BadRequestException("toTeamId must be different from current team.");
        }

        player.team = toTeam;
        if (request.uniformNumber() != null) {
            player.uniformNumber = normalizeUniformNumber(request.uniformNumber());
        }
        repository.persistAndFlush(player);
        recordTransferHistory(player, fromTeam, toTeam, request.transferredAt());
        return Response.ok(player).build();
    }

    @GET
    @APIResponseSchema(value = Player[].class, responseCode = "200")
    public Response list(
        @QueryParam("include_deleted") Boolean includeDeleted,
        @QueryParam("team_id") Integer teamId,
        @QueryParam("position") String position,
        @QueryParam("name") String name
    ) {
        var players = repository.search(includeDeleted, teamId, position, name);
        return Response.ok(players).build();
    }

    @POST
    @Transactional
    @Path("/")
    @APIResponses({
        @APIResponse(responseCode = "201"),
        @APIResponse(responseCode = "404"),
        @APIResponse(responseCode = "409", content = @Content(schema = @Schema(implementation = ErrorMessage.class))),
        @APIResponse(responseCode = "500")
    })
    public Response create(PlayerCreateRequest request) {
        if (request == null
            || request.teamId() == null
            || request.name() == null || request.name().isBlank()
            || request.uniformNumber() == null
            || request.position() == null || request.position().isBlank()) {
            throw new BadRequestException("teamId, name, uniformNumber and position are required.");
        }

        Team team = teamRepository.findByIdOptional(request.teamId()).orElseThrow(NotFoundException::new);
        Player player = new Player();
        player.team = team;
        player.name = normalizeName(request.name());
        player.uniformNumber = normalizeUniformNumber(request.uniformNumber());
        player.position = normalizePosition(request.position());
        repository.persistAndFlush(player);
        recordTransferHistory(player, null, player.team, player.createdAt);
        return Response.status(Status.CREATED).entity(player).build();
    }

    @PUT
    @Transactional
    @Path("/{id}")
    @APIResponses({
        @APIResponse(responseCode = "200"),
        @APIResponse(responseCode = "400"),
        @APIResponse(responseCode = "404"),
        @APIResponse(responseCode = "409", content = @Content(schema = @Schema(implementation = ErrorMessage.class))),
        @APIResponse(responseCode = "500")
    })
    public Response update(@PathParam("id") Integer playerId, PlayerUpdateRequest request) {
        if (request == null) {
            throw new BadRequestException("request body is required.");
        }

        if (request.teamId() == null
            && request.name() == null
            && request.uniformNumber() == null
            && request.position() == null) {
            throw new BadRequestException("At least one updatable field is required.");
        }

        Player player = repository.findByIdOptional(playerId).orElseThrow(NotFoundException::new);
        Team previousTeam = player.team;
        Team nextTeam = previousTeam;
        boolean teamChanged = false;

        if (request.teamId() != null) {
            nextTeam = teamRepository.findByIdOptional(request.teamId()).orElseThrow(NotFoundException::new);
            teamChanged = !nextTeam.id.equals(previousTeam.id);
            player.team = nextTeam;
        }

        if (request.name() != null) {
            player.name = normalizeName(request.name());
        }

        if (request.uniformNumber() != null) {
            player.uniformNumber = normalizeUniformNumber(request.uniformNumber());
        }

        if (request.position() != null) {
            player.position = normalizePosition(request.position());
        }

        repository.persistAndFlush(player);
        if (teamChanged) {
            recordTransferHistory(player, previousTeam, nextTeam, LocalDateTime.now());
        }
        return Response.ok(player).build();
    }

    @DELETE
    @Transactional
    @Path("/{id}")
    public Response delete(@PathParam("id") Integer playerId) {
        repository.findByIdOptional(playerId).ifPresentOrElse(player -> {
            player.isActive = false;
            repository.persist(player);
        }, () -> {
            throw new NotFoundException();
        });
        return Response.status(Status.NO_CONTENT).build();
    }

    private String normalizeName(String name) {
        String normalized = name.trim();
        if (normalized.isBlank()) {
            throw new BadRequestException("name must not be blank.");
        }

        if (normalized.length() > PLAYER_NAME_MAX_LENGTH) {
            throw new BadRequestException("name must be 64 characters or less.");
        }
        return normalized;
    }

    private Integer normalizeUniformNumber(Integer uniformNumber) {
        if (uniformNumber < MIN_UNIFORM_NUMBER || uniformNumber > MAX_UNIFORM_NUMBER) {
            throw new BadRequestException("uniformNumber must be between 0 and 999.");
        }
        return uniformNumber;
    }

    private String normalizePosition(String position) {
        String normalized = position.trim().toUpperCase();
        if (!SUPPORTED_POSITIONS.contains(normalized)) {
            throw new BadRequestException("position must be one of P, C, IF, OF.");
        }
        return normalized;
    }

    private void recordTransferHistory(Player player, Team fromTeam, Team toTeam, LocalDateTime transferredAt) {
        PlayerTransferHistory history = new PlayerTransferHistory();
        history.player = player;
        history.fromTeam = fromTeam;
        history.toTeam = toTeam;
        history.transferredAt = transferredAt == null ? LocalDateTime.now() : transferredAt;
        transferHistoryRepository.persist(history);
    }

    private LocalDateTime parseDateTimeOrNull(String value, String fieldName) {
        if (value == null || value.isBlank()) {
            return null;
        }
        try {
            return LocalDateTime.parse(value.trim());
        } catch (DateTimeParseException ex) {
            throw new BadRequestException(fieldName + " must be ISO-8601 date-time format.");
        }
    }

    private int validateTransferHistoryLimit(Integer limit) {
        if (limit == null) {
            return DEFAULT_TRANSFER_HISTORY_LIMIT;
        }
        if (limit <= 0 || limit > MAX_TRANSFER_HISTORY_LIMIT) {
            throw new BadRequestException("limit must be between 1 and 500.");
        }
        return limit;
    }

    public record PlayerCreateRequest(
        Integer teamId,
        String name,
        Integer uniformNumber,
        String position
    ) {
    }

    public record PlayerUpdateRequest(
        Integer teamId,
        String name,
        Integer uniformNumber,
        String position
    ) {
    }

    public record PlayerTransferHistoryResponse(
        Integer id,
        Integer playerId,
        Integer fromTeamId,
        Integer toTeamId,
        LocalDateTime transferredAt,
        LocalDateTime createdAt
    ) {
    }

    public record PlayerTransferRequest(
        Integer toTeamId,
        Integer uniformNumber,
        LocalDateTime transferredAt
    ) {
    }
}