# src/main/resources/openapi-definition/openapi.yaml
---
openapi: 3.1.0
info:
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: JPA Sample Application
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Local server
components:
  schemas:
    ErrorMessage:
      type: object
      properties:
        message:
          type: string
        detailMessage:
          type: string
    LocalDateTime:
      type: string
      format: date-time
      examples:
      - 2022-03-10T12:15:50
    Manager:
      type: object
      properties:
        id:
          type: integer
          format: int32
        team:
          $ref: "#/components/schemas/Team"
        name:
          type: string
        createdAt:
          $ref: "#/components/schemas/LocalDateTime"
        updatedAt:
          $ref: "#/components/schemas/LocalDateTime"
    Player:
      type: object
      properties:
        id:
          type: integer
          format: int32
        team:
          $ref: "#/components/schemas/Team"
        name:
          type: string
        uniformNumber:
          type: integer
          format: int32
        position:
          type: string
        createdAt:
          $ref: "#/components/schemas/LocalDateTime"
        updatedAt:
          $ref: "#/components/schemas/LocalDateTime"
    PlayerCreateRequest:
      type: object
      properties:
        teamId:
          type: integer
          format: int32
        name:
          type: string
        uniformNumber:
          type: integer
          format: int32
        position:
          type: string
    PlayerTransferHistoryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        playerId:
          type: integer
          format: int32
        fromTeamId:
          type: integer
          format: int32
        toTeamId:
          type: integer
          format: int32
        transferredAt:
          $ref: "#/components/schemas/LocalDateTime"
        createdAt:
          $ref: "#/components/schemas/LocalDateTime"
    PlayerTransferRequest:
      type: object
      properties:
        toTeamId:
          type: integer
          format: int32
        uniformNumber:
          type: integer
          format: int32
        transferredAt:
          $ref: "#/components/schemas/LocalDateTime"
    PlayerUpdateRequest:
      type: object
      properties:
        teamId:
          type: integer
          format: int32
        name:
          type: string
        uniformNumber:
          type: integer
          format: int32
        position:
          type: string
    Team:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        createdAt:
          $ref: "#/components/schemas/LocalDateTime"
        updatedAt:
          $ref: "#/components/schemas/LocalDateTime"
    TeamCreateRequest:
      type: object
      properties:
        name:
          type: string
paths:
  /v1/manager:
    get:
      parameters:
      - name: include_deleted
        in: query
        schema:
          type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Manager"
      summary: List
      tags:
      - Manager Resource
  /v1/manager/{id}:
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Manager"
      summary: Get By Id
      tags:
      - Manager Resource
  /v1/panache-sample/teams:
    get:
      parameters:
      - name: name
        in: query
        schema:
          type: string
      - name: page
        in: query
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        schema:
          type: integer
          format: int32
          default: 5
      responses:
        "200":
          description: OK
      summary: List Teams By Panache
      tags:
      - Panache Sample Resource
  /v1/panache-sample/teams/native-cte-stats:
    get:
      parameters:
      - name: min_manager_count
        in: query
        schema:
          type: integer
          format: int32
          default: 0
      responses:
        "200":
          description: OK
      summary: Team Stats By Native Cte
      tags:
      - Panache Sample Resource
  /v1/panache-sample/teams/native-stats:
    get:
      responses:
        "200":
          description: OK
      summary: Team Stats By Native Query
      tags:
      - Panache Sample Resource
  /v1/panache-sample/teams/native-summary:
    get:
      responses:
        "200":
          description: OK
      summary: Team Summary By Native Query
      tags:
      - Panache Sample Resource
  /v1/player:
    get:
      parameters:
      - name: include_deleted
        in: query
        schema:
          type: boolean
      - name: name
        in: query
        schema:
          type: string
      - name: position
        in: query
        schema:
          type: string
      - name: team_id
        in: query
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Player"
      summary: List
      tags:
      - Player Resource
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerCreateRequest"
        required: true
      responses:
        "201": {}
        "404": {}
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "500": {}
        "400":
          description: Bad Request
      summary: Create
      tags:
      - Player Resource
  /v1/player/{id}:
    put:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerUpdateRequest"
        required: true
      responses:
        "200":
          description: OK
        "400": {}
        "404": {}
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "500": {}
      summary: Update
      tags:
      - Player Resource
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
      summary: Get By Id
      tags:
      - Player Resource
    delete:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
      summary: Delete
      tags:
      - Player Resource
  /v1/player/{id}/transfers:
    post:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerTransferRequest"
        required: true
      responses:
        "200":
          description: OK
        "400": {}
        "404": {}
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "500": {}
      summary: Transfer
      tags:
      - Player Resource
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: from
        in: query
        schema:
          type: string
      - name: limit
        in: query
        schema:
          type: integer
          format: int32
      - name: to
        in: query
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlayerTransferHistoryResponse"
      summary: Get Transfers
      tags:
      - Player Resource
  /v1/team:
    get:
      parameters:
      - name: include_deleted
        in: query
        schema:
          type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"
      summary: List
      tags:
      - Team Resource
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamCreateRequest"
        required: true
      responses:
        "201": {}
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "500": {}
        "400":
          description: Bad Request
      summary: Create
      tags:
      - Team Resource
  /v1/team/{id}:
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
      summary: Get By Id
      tags:
      - Team Resource
    delete:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
      summary: Delete
      tags:
      - Team Resource